Check in `Cargo.lock`
It's best practice to check it in for crates with executables:
https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html

Other than it being "the right thing to do", I am currently trying to
get Actiserve built using Nix, which is much easier if the lock file is
checked in. I'm doing this, because Nix is a very easy way to get a
Mastodon server up and experiment with the Mastodon/relay protocol.